modules/location-producer/requirements.txt
GeoAlchemy2==0.8.4
jsonschema==3.0.2
grpcio-tools==1.51.1
kafka-python==2.0.1
SQLAlchemy==1.3.19